change(web): use function to apply browserstack-specific settings to standard test-runner config
